requirements.txt
Django>=5.0,<5.1
psycopg2-binary>=2.9
django-unfold>=0.18.0
django-import-export>=3.3.0
